clean up

+++ torch_xla/csrc/xla_custom_call.cpp
@@ -6,8 +6,8 @@
 // hardcoded. Need to be sent in the args. 
 // TODO: template the datatype?
 void argmin_custom(void* out, const void** in) {
-  int* out_buf = reinterpret_cast<int*>(out);
-  const int* in0 = reinterpret_cast<const int*>(in[0]);
+  float* out_buf = reinterpret_cast<float*>(out);
+  const float* in0 = reinterpret_cast<const float*>(in[0]);
   int min_idx = 0;
   std::cout << "Custom call entered - Assumes only 4 contiguous elements" << std::endl;
   for (int i = 0; i < 4; ++i) {
